Changed from float to double for better performance and accuracy.

+++ core/src/main/java/com/physics/math/ComplexNumber.java
@@ -1,10 +1,10 @@
 package com.physics.math;
 
 public class ComplexNumber {
-    private float real;
-    private float imag;
+    private double real;
+    private double imag;
 
-    public ComplexNumber(float real, float imag) {
+    public ComplexNumber(double real, double imag) {
         this.real = real;
         this.imag = imag;
     }
@@ -13,30 +13,30 @@ public class ComplexNumber {
         this.imag = 0;
     }
 
-    public void add(float a, float b) {
+    public void add(double a, double b) {
         this.real += a;
         this.imag += b;
     }
 
-    public void subtract(float a, float b) {
+    public void subtract(double a, double b) {
         this.real -= a;
         this.imag -= b;
     }
 
     public void multiply(ComplexNumber c) {
-        float newReal = this.real * c.real - this.imag * c.imag;
-        float newImag = this.real * c.imag + this.imag * c.real;
+        double newReal = this.real * c.real - this.imag * c.imag;
+        double newImag = this.real * c.imag + this.imag * c.real;
         this.real = newReal;
         this.imag = newImag;
     }
 
     public void divide(ComplexNumber c) {
-        float denom = c.real * c.real + c.imag * c.imag;
+        double denom = c.real * c.real + c.imag * c.imag;
         if (denom == 0) {
             throw new ArithmeticException("Division durch Null");
         }
-        float newReal = (this.real * c.real + this.imag * c.imag) / denom;
-        float newImag = (this.imag * c.real - this.real * c.imag) / denom;
+        double newReal = (this.real * c.real + this.imag * c.imag) / denom;
+        double newImag = (this.imag * c.real - this.real * c.imag) / denom;
         this.real = newReal;
         this.imag = newImag;
     }
@@ -45,19 +45,19 @@ public class ComplexNumber {
         return new ComplexNumber(real, -imag);
     }
 
-    public float getReal() {
+    public double getReal() {
         return real;
     }
 
-    public void setReal(float real) {
+    public void setReal(double real) {
         this.real = real;
     }
 
-    public float getImag() {
+    public double getImag() {
         return imag;
     }
 
-    public void setImag(float imag) {
+    public void setImag(double imag) {
         this.imag = imag;
     }
